Add notes2actions.py under the CRAPL license.
This is an old (and unused) file.  It is basically what Vivi (the
Virtual Violinist) replaces.  This is really bad code, and nobody
should look at this and think that it's anything other than an
embarrassment to the Percival name.

However, on the off chance that it might be of use to
somebody -- if only to illustrate what Vivi does, or what kinds of
hard-coding you need to do if you don't use Vivi, or something
like that -- I'm adding it to the project.

Be warned.