Improve robustness of build system.
That's not to suggest that the build system _is_ robust; in fact,
the more I learn about scons, the less I like it.

But you can still "improve robustness" of scons in the same way
that electing the Liberals instead of the Conservatives into power
during #elxn41 would "improve trustworthiness" of the Canadian
governement.

I'm undecided between my long-term strategy of always voting for
the third party (meaning the NDP) to show how stupid "first past
the post" is, or voting for the Liberals just to get rid of the
Conservatives.  I don't care[1] about finances, military, the
environment or any of the traditional conservative (or
anti-conservative) issues; I'm simply digusted by their complete
contempt for democracy and openness.

[1] well, actually I _do_ care about some of those -- but my
concern with democracy itself far outweighs my concerns about any
particular policies of any particular party.