Updated model.
Graham's note: I made some minor changes to the .blend file
(doing some parenting and applying rotations to make the python
scripts work), and screwed up the bow.  The bow is supposed to be
mirrored, but I accidently broke it, so at the moment we only have
half a bow.