Build: add option . to disable fftw_cleanup()
Vivi uses aubio, which uses fftw, which makes really annoying use
of global variables for plans.  These are cleaned up by running
fftw_cleanup(), but it must be run once and only once.
Artifastring normally runs fftw_cleanup(), but when Vivi calls
aubio and deletes the aubio object, it *also* tries to run
fftw_cleanup(), which then borks.

Solution: a configure option for artifastring which disables the
fftw_cleanup() inside Artifatring.  This should only be used if
Artifastring is to be used in conjunction with another library
which insists on calling fftw_cleanup() itself.